Support more complex scenarios
A new setting was added to allow for the activation of CORS for certain URL patterns,
regardless of the request origin. This can be combined with the current available
settings to support more complex authorization scenarios.

Documentation and tests have been also included to give more insight into the setting.